add failing test for review_inactive to book_spec
OK, so simplecov is showing 2 untested lines:
app/controllers/reviews_controller.rb:18-19
which redirect if the book is inactive.
Interactive testing in the browser shows this works as expected.

It is my test that is incorrect.
spec/features/books_spec.rb:59
tries to visit the page and click "New", but capybara is not seeing the
"New" button.
Can someone pull this locally and figure out the test?