make sort start to work
pull ou the fields that I have not figured out how to sort.
This does not have complete coverage.
The sort problem is: we are using an ActiveRecord sort (Book.order) -
which doesn't let us sort by foreign fields. To do this, we'll either
need a join, or to sort again *after* pulling in from ActiveRecord.
Neither solution is likely to be pretty, and I haven't figured how to
factor it out.
I think this is another reason we might wish to roll-our-own ratings -
have average rating be a Book field.